+++ packages/fuselage/src/components/Sidebar/SidebarDivider.tsx
@@ -1,10 +1,5 @@
 import { Divider } from '../Divider';
 
-/**
- * @deprecated Use `SidebarV2Divider` instead.
- */
-const SidebarDivider = () => (
+export const SidebarDivider = () => (
   <Divider rcx-sidebar--divider marginBlockStart={-2} marginBlockEnd={0} />
 );
-
-export default SidebarDivider;