+++ packages/fuselage/src/components/SidebarV2/SidebarDivider.tsx
@@ -1,5 +1,0 @@
-import { Divider } from '../Divider';
-
-export const SidebarDivider = () => (
-  <Divider rcx-sidebar-v2--divider marginBlockStart={-2} marginBlockEnd={0} />
-);