+++ packages/fuselage/src/components/SidebarV2/SidebarFooter/index.ts
@@ -1,2 +1,0 @@
-export { SidebarFooter as SidebarV2Footer } from './SidebarFooter';
-export { SidebarFooterContent as SidebarV2FooterContent } from './SidebarFooterContent';