+++ packages/fuselage/src/components/SidebarV2/SidebarMedia/index.ts
@@ -1,3 +1,0 @@
-export { SidebarMedia as SidebarV2Media } from './SidebarMedia';
-export { SidebarMediaController as SidebarV2MediaController } from './SidebarMediaController';
-export { SidebarMediaTitle as SidebarV2MediaTitle } from './SidebarMediaTitle';